Mejoras estéticas y arreglos de bugs varios

+++ app/src/main/java/com/franciscolinares/ubb/estadistica/MVPFragment.kt
@@ -132,7 +132,7 @@ class MVPFragment : Fragment() {
         val mejoresJugadores = ArrayList(jugadoresOrdenados.take(cantidad))
         for (j in mejoresJugadores) {
             val dato = JugadorEstadistica(
-                j["minutos"].toString().toFloat(),
+                j["minutos"].toString(),
                 j["dorsal"].toString(),
                 j["nombre"].toString(),
                 j["equipo"].toString(),
